requirements.txt
qiskit
pennylane
amazon-braket-pennylane-plugin
pennylane-qiskit